refactor(search page): move search input to header and use query params
Refactored the search page: the search input is now located in the header instead of the search page itself. When the user submits a query, it is passed to the search page via URL query parameters. The search page then fetches and displays the results based on the provided query.